requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
supabase
flask
pyspellchecker[fr]
